Simplifying SweepChunk class. Removing unnecessary sweep includes in code outside of modules/linear_boltzmann_solvers.
