+++ framework/graphs/directed_graph_vertex.h
@@ -3,6 +3,7 @@
 #include "framework/graphs/graph.h"
 
 #include <map>
+#include <set>
 
 namespace opensn
 {
