+++ framework/math/spatial_discretization/spatial_discretization.h
@@ -11,6 +11,7 @@
 
 #include <vector>
 #include <map>
+#include <set>
 
 namespace opensn
 {
