+++ framework/mesh/raytrace/raytracer.cc
@@ -2,7 +2,9 @@
 #include "framework/mesh/cell/cell.h"
 #include "framework/mesh/mesh_continuum/mesh_continuum.h"
 #include "framework/logging/log.h"
+
 #include <algorithm>
+#include <set>
 
 namespace opensn
 {
